+++ ThirdPartyAdapters/unity/unity/src/main/java/com/google/ads/mediation/unity/UnityMediationAdapter.java
@@ -44,6 +44,7 @@ import com.google.android.gms.ads.mediation.rtb.RtbAdapter;
 import com.google.android.gms.ads.mediation.rtb.RtbSignalData;
 import com.google.android.gms.ads.mediation.rtb.SignalCallbacks;
 import com.unity3d.ads.IUnityAdsInitializationListener;
+import com.unity3d.ads.TokenConfiguration;
 import com.unity3d.ads.UnityAds;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -156,6 +157,9 @@ public class UnityMediationAdapter extends RtbAdapter {
   static final String KEY_PLACEMENT_ID = "zoneId";
 
   static final String KEY_WATERMARK = "watermark";
+
+  private final UnityAdsWrapper unityAdsWrapper;
+
   private final UnityInitializer unityInitializer;
 
   private final UnityBannerViewFactory unityBannerViewFactory;
@@ -184,6 +188,7 @@ public class UnityMediationAdapter extends RtbAdapter {
 
   public UnityMediationAdapter() {
     unityInitializer = UnityInitializer.getInstance();
+    unityAdsWrapper = new UnityAdsWrapper();
     unityBannerViewFactory = new UnityBannerViewFactory();
     this.unityAdsLoader = new UnityAdsLoader();
   }
@@ -191,11 +196,13 @@ public class UnityMediationAdapter extends RtbAdapter {
   @Override
   public void collectSignals(
       @NonNull RtbSignalData rtbSignalData, @NonNull SignalCallbacks signalCallbacks) {
+    AdFormat adFormat = getAdFormat(rtbSignalData);
+    com.unity3d.ads.AdFormat unityAdFormat = null;
+
     // For banner ad format, Unity Ads SDK requires an activity context to load the banner ad. So,
     // fail here so that Unity bidder will not bid if the ad request was made with a non-activity
     // context.
-    if (getAdFormat(rtbSignalData) == AdFormat.BANNER
-        && !(rtbSignalData.getContext() instanceof Activity)) {
+    if (adFormat == AdFormat.BANNER && !(rtbSignalData.getContext() instanceof Activity)) {
       signalCallbacks.onFailure(
           new AdError(
               ERROR_CONTEXT_NOT_ACTIVITY,
@@ -204,21 +211,45 @@ public class UnityMediationAdapter extends RtbAdapter {
       return;
     }
 
-    UnityAds.getToken(
-        token -> {
-          if (token == null) {
-            token = "";
-          }
-          signalCallbacks.onSuccess(token);
-        });
+    if (adFormat == AdFormat.BANNER) {
+      unityAdFormat = com.unity3d.ads.AdFormat.BANNER;
+    } else if (adFormat == AdFormat.REWARDED || adFormat == AdFormat.REWARDED_INTERSTITIAL) {
+      unityAdFormat = com.unity3d.ads.AdFormat.REWARDED;
+    } else if (adFormat == AdFormat.INTERSTITIAL) {
+      unityAdFormat = com.unity3d.ads.AdFormat.INTERSTITIAL;
+    } else {
+      Log.w(TAG, "Unsupported ad format for Unity Ads: " + adFormat);
+    }
+
+    if (unityAdFormat != null) {
+      TokenConfiguration tokenConfiguration = new TokenConfiguration(unityAdFormat);
+      unityAdsWrapper.getToken(
+          tokenConfiguration,
+          token -> {
+            if (token == null) {
+              token = "";
+            }
+            signalCallbacks.onSuccess(token);
+          });
+    } else {
+      unityAdsWrapper.getToken(
+          token -> {
+            if (token == null) {
+              token = "";
+            }
+            signalCallbacks.onSuccess(token);
+          });
+    }
   }
 
   @VisibleForTesting
   UnityMediationAdapter(
       UnityInitializer unityInitializer,
+      UnityAdsWrapper unityAdsWrapper,
       UnityBannerViewFactory unityBannerViewFactory,
       UnityAdsLoader unityAdsLoader) {
     this.unityInitializer = unityInitializer;
+    this.unityAdsWrapper = unityAdsWrapper;
     this.unityBannerViewFactory = unityBannerViewFactory;
     this.unityAdsLoader = unityAdsLoader;
   }