+++ ThirdPartyAdapters/unity/unity/src/test/kotlin/com/google/ads/mediation/unity/UnityMediationAdapterTest.kt
@@ -40,10 +40,13 @@ import com.google.android.gms.ads.mediation.rtb.RtbSignalData
 import com.google.android.gms.ads.mediation.rtb.SignalCallbacks
 import com.google.common.truth.Truth.assertThat
 import com.unity3d.ads.IUnityAdsInitializationListener
+import com.unity3d.ads.IUnityAdsTokenListener
+import com.unity3d.ads.TokenConfiguration
 import com.unity3d.ads.UnityAds.UnityAdsInitializationError
 import com.unity3d.ads.UnityAdsLoadOptions
 import com.unity3d.ads.metadata.MediationMetaData
 import com.unity3d.services.banners.UnityBannerSize
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -99,7 +102,12 @@ class UnityMediationAdapterTest {
   @Before
   fun setUp() {
     unityMediationAdapter =
-      UnityMediationAdapter(unityInitializer, unityBannerViewFactory, unityAdsLoader)
+      UnityMediationAdapter(
+        unityInitializer,
+        unityAdsWrapper,
+        unityBannerViewFactory,
+        unityAdsLoader,
+      )
     mediationConfigurations = emptyList()
 
     adSize = AdSize.BANNER
@@ -224,6 +232,131 @@ class UnityMediationAdapterTest {
     val adError = adErrorCaptor.firstValue
     assertThat(adError.code).isEqualTo(ERROR_CONTEXT_NOT_ACTIVITY)
     assertThat(adError.domain).isEqualTo(ADAPTER_ERROR_DOMAIN)
+    verifyNoMoreInteractions(signalCallbacks)
+  }
+
+  @Test
+  fun collectSignals_forBannerFormatAndActivityContext_invokesSignalCallbacks() {
+    whenever(unityAdsWrapper.getToken(any(), any())) doAnswer
+      { invocation ->
+        val callback = invocation.arguments[1] as IUnityAdsTokenListener
+        callback.onUnityAdsTokenReady(TEST_TOKEN)
+      }
+
+    val rtbSignalData =
+      RtbSignalData(
+        activity,
+        listOf(MediationConfiguration(AdFormat.BANNER, /* serverParameters= */ bundleOf())),
+        /* networkExtras= */ bundleOf(),
+        AdSize.BANNER,
+      )
+
+    unityMediationAdapter.collectSignals(rtbSignalData, signalCallbacks)
+
+    verify(signalCallbacks).onSuccess(any())
+    val tokenConfigCaptor = argumentCaptor<TokenConfiguration>()
+    verify(unityAdsWrapper).getToken(tokenConfigCaptor.capture(), any())
+    assertEquals(com.unity3d.ads.AdFormat.BANNER, tokenConfigCaptor.firstValue.adFormat)
+    verifyNoMoreInteractions(signalCallbacks)
+  }
+
+  @Test
+  fun collectSignals_forInterstitialFormat_invokesSignalCallbacks() {
+    whenever(unityAdsWrapper.getToken(any(), any())) doAnswer
+      { invocation ->
+        val callback = invocation.arguments[1] as IUnityAdsTokenListener
+        callback.onUnityAdsTokenReady(TEST_TOKEN)
+      }
+
+    val rtbSignalData =
+      RtbSignalData(
+        activity,
+        listOf(MediationConfiguration(AdFormat.INTERSTITIAL, /* serverParameters= */ bundleOf())),
+        /* networkExtras= */ bundleOf(),
+        null,
+      )
+
+    unityMediationAdapter.collectSignals(rtbSignalData, signalCallbacks)
+
+    verify(signalCallbacks).onSuccess(any())
+    val tokenConfigCaptor = argumentCaptor<TokenConfiguration>()
+    verify(unityAdsWrapper).getToken(tokenConfigCaptor.capture(), any())
+    assertEquals(com.unity3d.ads.AdFormat.INTERSTITIAL, tokenConfigCaptor.firstValue.adFormat)
+    verifyNoMoreInteractions(signalCallbacks)
+  }
+
+  @Test
+  fun collectSignals_forRewardedFormat_invokesSignalCallbacks() {
+    whenever(unityAdsWrapper.getToken(any(), any())) doAnswer
+      { invocation ->
+        val callback = invocation.arguments[1] as IUnityAdsTokenListener
+        callback.onUnityAdsTokenReady(TEST_TOKEN)
+      }
+
+    val rtbSignalData =
+      RtbSignalData(
+        activity,
+        listOf(MediationConfiguration(AdFormat.REWARDED, /* serverParameters= */ bundleOf())),
+        /* networkExtras= */ bundleOf(),
+        null,
+      )
+
+    unityMediationAdapter.collectSignals(rtbSignalData, signalCallbacks)
+
+    verify(signalCallbacks).onSuccess(any())
+    val tokenConfigCaptor = argumentCaptor<TokenConfiguration>()
+    verify(unityAdsWrapper).getToken(tokenConfigCaptor.capture(), any())
+    assertEquals(com.unity3d.ads.AdFormat.REWARDED, tokenConfigCaptor.firstValue.adFormat)
+    verifyNoMoreInteractions(signalCallbacks)
+  }
+
+  @Test
+  fun collectSignals_forRewardedInterstitialFormat_invokesSignalCallbacks() {
+    whenever(unityAdsWrapper.getToken(any(), any())) doAnswer
+      { invocation ->
+        val callback = invocation.arguments[1] as IUnityAdsTokenListener
+        callback.onUnityAdsTokenReady(TEST_TOKEN)
+      }
+
+    val rtbSignalData =
+      RtbSignalData(
+        activity,
+        listOf(
+          MediationConfiguration(AdFormat.REWARDED_INTERSTITIAL, /* serverParameters= */ bundleOf())
+        ),
+        /* networkExtras= */ bundleOf(),
+        null,
+      )
+
+    unityMediationAdapter.collectSignals(rtbSignalData, signalCallbacks)
+
+    verify(signalCallbacks).onSuccess(any())
+    val tokenConfigCaptor = argumentCaptor<TokenConfiguration>()
+    verify(unityAdsWrapper).getToken(tokenConfigCaptor.capture(), any())
+    // GMA's rewarded interstitial format is mapped to Unity Ads's rewarded format.
+    assertEquals(com.unity3d.ads.AdFormat.REWARDED, tokenConfigCaptor.firstValue.adFormat)
+    verifyNoMoreInteractions(signalCallbacks)
+  }
+
+  @Test
+  fun collectSignals_forUnsupportedFormat_invokesSignalCallbacks() {
+    whenever(unityAdsWrapper.getToken(any())) doAnswer
+      { invocation ->
+        val callback = invocation.arguments[0] as IUnityAdsTokenListener
+        callback.onUnityAdsTokenReady(TEST_TOKEN)
+      }
+
+    val rtbSignalData =
+      RtbSignalData(
+        activity,
+        listOf(MediationConfiguration(AdFormat.NATIVE, /* serverParameters= */ bundleOf())),
+        /* networkExtras= */ bundleOf(),
+        null,
+      )
+
+    unityMediationAdapter.collectSignals(rtbSignalData, signalCallbacks)
+    verify(unityAdsWrapper).getToken(any())
+    verify(signalCallbacks).onSuccess(any())
     verifyNoMoreInteractions(signalCallbacks)
   }
 
@@ -829,5 +962,6 @@ class UnityMediationAdapterTest {
     const val TEST_VERSION_NUMBER = "100"
     const val TEST_GAME_ID = "gameId"
     const val TEST_PLACEMENT_ID = "placementId"
+    const val TEST_TOKEN = "test_token"
   }
 }